requirements.txt
tensorboardX==2.1
pyyaml==3.12
lpips==0.1.3
tqdm==4.61.1
matplotlib==3.4.2
sklearn
opencv-python
latex